hacky throwaway version of data perm graph set
I just wrote this to make `mt/with-data-permissions-for-all-users!`
work, so existing tests didn't need to be rewritten. I didn't worry at
all about... lots of things! For example, it assumes that every graph it
receives will include native permissions, though this isn't guaranteed
to be true (though I think the frontend does in fact include it every
time).